+++ src/wrapper/settings.rs
@@ -14,6 +14,7 @@ struct SecureSettings {
     storage_dir: Option<Path>,
     cache_dir: Option<Path>,
     version_check: Option<bool>,
+    ubuntu_mirror: Option<String>,
     site_settings: BTreeMap<Path, SecureSettings>,
 }
 
@@ -36,6 +37,9 @@ pub fn secure_settings_validator<'a>(has_children: bool)
         ("version_check".to_string(), Box::new(V::Scalar {
             optional: true,
             .. Default::default()}) as Box<V::Validator>),
+        ("ubuntu_mirror".to_string(), Box::new(V::Scalar {
+            optional: true,
+            .. Default::default()}) as Box<V::Validator>),
     );
     if has_children {
         members.push(
@@ -54,6 +58,7 @@ pub fn secure_settings_validator<'a>(has_children: bool)
 struct InsecureSettings {
     version_check: Option<bool>,
     shared_cache: Option<bool>,
+    ubuntu_mirror: Option<String>,
 }
 
 pub fn insecure_settings_validator<'a>() -> Box<V::Validator + 'a> {
@@ -64,6 +69,9 @@ pub fn insecure_settings_validator<'a>() -> Box<V::Validator + 'a> {
         ("shared_cache".to_string(), Box::new(V::Scalar {
             optional: true,
             .. Default::default()}) as Box<V::Validator + 'a>),
+        ("ubuntu_mirror".to_string(), Box::new(V::Scalar {
+            optional: true,
+            .. Default::default()}) as Box<V::Validator>),
     ), .. Default::default()}) as Box<V::Validator>;
 }
 
@@ -88,6 +96,7 @@ pub fn read_settings(project_root: &Path)
     let mut int_settings = Settings {
         version_check: true,
         uid_map: None,
+        ubuntu_mirror: "mirror://mirrors.ubuntu.com/mirrors.txt".to_string(),
     };
     let mut secure_files = vec!();
     if let Some(home) = getenv("VAGGA_USER_HOME") {
@@ -115,6 +124,9 @@ pub fn read_settings(project_root: &Path)
         if let Some(val) = cfg.version_check {
             int_settings.version_check = val;
         }
+        if let Some(ref val) = cfg.ubuntu_mirror {
+            int_settings.ubuntu_mirror = val.clone();
+        }
         if let Some(cfg) = cfg.site_settings.get(project_root) {
             for (k, v) in cfg.allowed_dirs.iter() {
                 ext_settings.allowed_dirs.insert(k.clone(), v.clone());
@@ -129,6 +141,9 @@ pub fn read_settings(project_root: &Path)
             if let Some(val) = cfg.version_check {
                 int_settings.version_check = val;
             }
+            if let Some(ref val) = cfg.ubuntu_mirror {
+                int_settings.ubuntu_mirror = val.clone();
+            }
         }
     }
     let mut insecure_files = vec!();
@@ -142,6 +157,9 @@ pub fn read_settings(project_root: &Path)
             &*insecure_settings_validator(), Default::default()));
         if let Some(val) = cfg.version_check {
             int_settings.version_check = val;
+        }
+        if let Some(ref val) = cfg.ubuntu_mirror {
+            int_settings.ubuntu_mirror = val.clone();
         }
         if let Some(val) = cfg.shared_cache {
             ext_settings.shared_cache = val;